# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for fleet management",
        "title": "Fleet Management API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/maintenance-logs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaintenanceLog"
                ],
                "summary": "Get all maintenance logs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.MaintenanceLog"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaintenanceLog"
                ],
                "summary": "Create maintenance log",
                "parameters": [
                    {
                        "description": "Maintenance Log",
                        "name": "log",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MaintenanceLog"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.MaintenanceLog"
                        }
                    }
                }
            }
        },
        "/maintenance-logs/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaintenanceLog"
                ],
                "summary": "Get maintenance log by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MaintenanceLog"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaintenanceLog"
                ],
                "summary": "Update maintenance log",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Maintenance Log",
                        "name": "log",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MaintenanceLog"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MaintenanceLog"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "MaintenanceLog"
                ],
                "summary": "Delete maintenance log",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "deleted",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/maintenance-request-items": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaintenanceRequestItem"
                ],
                "summary": "Get all maintenance request items",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.MaintenanceRequestItem"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaintenanceRequestItem"
                ],
                "summary": "Create maintenance request item",
                "parameters": [
                    {
                        "description": "Item",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MaintenanceRequestItem"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.MaintenanceRequestItem"
                        }
                    }
                }
            }
        },
        "/maintenance-request-items/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaintenanceRequestItem"
                ],
                "summary": "Get maintenance request item by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MaintenanceRequestItem"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaintenanceRequestItem"
                ],
                "summary": "Update maintenance request item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Item",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MaintenanceRequestItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MaintenanceRequestItem"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "MaintenanceRequestItem"
                ],
                "summary": "Delete maintenance request item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "deleted",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.MaintenanceLog": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "created, approved, in_progress, done, cancelled",
                    "type": "string"
                },
                "cost": {
                    "type": "number"
                },
                "createdBy": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "maintenanceRequestId": {
                    "type": "integer"
                }
            }
        },
...
        "models.MaintenanceRequestItem": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "number"
                },
                "itemName": {
                    "type": "string"
                },
                "maintenanceRequestId": {
                    "type": "integer"
                },
                "note": {
                    "type": "string"
                },
                "status": {
                    "description": "pending, done, cancelled",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}